+++ doc/changes.rst
@@ -12,10 +12,13 @@ Change Log
 * Fixed bugs that could cause bricks to be slightly too big, and that
   incorrectly special-cased the north pole with brick sizes that don't
   evenly divide 180 degrees (PR `#84`_).
+* Added tests of :mod:`desiutil.sklearn` (PR `#86`_).
 
 .. _`#80`: https://github.com/desihub/desiutil/pull/80
 .. _`#81`: https://github.com/desihub/desiutil/pull/81
 .. _`#84`: https://github.com/desihub/desiutil/pull/84
+.. _`#86`: https://github.com/desihub/desiutil/pull/86
+
 
 1.9.6 (2017-07-12)
 ------------------